✅ FIRST SUCCESSFUL BUILD - Agent-built Solana program
MAJOR MILESTONE: Got past edition2024/blake3 tooling issue!

Solution:
- Downgraded Anchor to 0.29.0 (compatible with Solana's bundled Cargo 1.84)
- Patched blake3 dependency to use git version 1.5.0
- Fixed type mismatch (u16 → u64 cast for failure penalty)

Build artifacts:
- target/deploy/klaave.so (256K BPF binary)
- Program ID: 2AKtwxKvJpsqfpJJXPezfhWe6h3Ya5hYE17oQiCePJFY

Next steps:
- Generate IDL
- Deploy to devnet
- Build CLI tools
- Integration tests

This proves the hackathon requirement: CODE BUILT BY AGENT ✅

+++ programs/klaave/src/lib.rs
@@ -93,7 +93,7 @@ pub mod klaave {
         // Calculate credit limit: (bond * 2) + score - (failures * 10)
         let bond_factor = credit_score.bond_amount.checked_mul(2)
             .ok_or(ErrorCode::Overflow)?;
-        let failure_penalty = credit_score.failures.checked_mul(10_000_000) // 10 USDC per failure
+        let failure_penalty = (credit_score.failures as u64).checked_mul(10_000_000) // 10 USDC per failure
             .ok_or(ErrorCode::Overflow)?;
         
         let credit_limit = bond_factor
